requirements.txt
streamlit>=1.30,<2
pandas>=2.0,<3
numpy>=1.23,<3
matplotlib>=3.7,<4
plotly>=5.0,<6

google-cloud-bigquery>=3.10,<4
google-auth>=2.0,<3
google-auth-oauthlib>=1.0,<2
google-cloud-bigquery-storage>=2.20,<3

pyarrow>=12,<16
db-dtypes>=1.2,<2
